requirements.txt
agno
google-genai
qdrant-client
pypdf
openai
python-dotenv
fastapi
uvicorn[standard]
resend
pydantic
pydantic-settings
groq
edge-tts
websockets
python-multipart
sqlalchemy
bcrypt
pyjwt